backend/requirements.txt
fastapi
uvicorn
solana
solders
pydantic
python-dotenv
beautifulsoup4
httpx
base58
construct
